requirements.txt
azure-functions==1.0.8
azure-storage==0.36.0